- move methods to toolbar component - remove duplicate methods and values from survey-filters, main-filters and search-form
