+++ apps/web-mzima-client/src/app/shared/components/toolbar/toolbar.component.html
@@ -13,12 +13,20 @@
     {{ pageTitle | translate }}
   </h1>
   <div class="toolbar__controls">
-    <app-search-form class="search-form" *ngIf="showSearchForm"></app-search-form>
+    <app-search-form
+      [isFiltersVisible]="isFiltersVisible"
+      [isMainFiltersOpen]="isMainFiltersOpen"
+      [form]="form"
+      (applyFilters)="applyFilters()"
+      (searchPosts)="searchPosts($event)"
+      class="search-form"
+      *ngIf="showSearchForm"
+    ></app-search-form>
     <mzima-client-button
       [fill]="isDesktop ? 'clear' : 'outline'"
       [color]="isDesktop ? 'gray' : 'secondary'"
       class="search-form__filters"
-      *ngIf="isDesktop || isMapView"
+      *ngIf="(isDesktop || isMapView) && showSearchForm"
       (buttonClick)="toggleFilters(!isFiltersVisible)"
       [data-qa]="'search-form__filters'"
     >
@@ -44,8 +52,53 @@
   >
     <mat-icon icon svgIcon="hamburger"></mat-icon>
   </mzima-client-button>
-  <app-main-filters></app-main-filters>
-  <app-survey-filters></app-survey-filters>
+  <app-main-filters
+    *ngIf="showSearchForm"
+    [form]="form"
+    [isFiltersVisible]="isFiltersVisible"
+    [canCreateSearch]="canCreateSearch"
+    [isMainFiltersOpen]="isMainFiltersOpen"
+    [isEditAvailable]="isEditAvailable"
+    [surveyList]="surveyList"
+    [savedSearches]="savedSearches"
+    [sources]="sources"
+    [statuses]="statuses"
+    [categoriesData]="categoriesData"
+    [activeSavedSearchValue]="activeSavedSearchValue"
+    (applyFilters)="applyFilters()"
+    (applyFilters)="clearFilter($event)"
+    (saveSearch)="saveSearch()"
+    (toggleFilters)="toggleFilters($event)"
+    (resetSavedFilter)="resetSavedFilter()"
+    (applyAndClose)="applyAndClose()"
+    (applySavedFilter)="applySavedFilter($event)"
+    (toggleSidebarFilters)="toggleSidebarFilters($event)"
+  ></app-main-filters>
+  <app-survey-filters
+    *ngIf="showSearchForm"
+    [form]="form"
+    [total]="total"
+    [notMappedPostsCount]="notMappedPostsCount"
+    [collectionInfo]="collectionInfo"
+    [surveyList]="surveyList"
+    [sources]="sources"
+    [isMapView]="isMapView"
+    [isLoggedIn]="isLoggedIn"
+    [isFiltersVisible]="isFiltersVisible"
+    [isMainFiltersOpen]="isMainFiltersOpen"
+    [isOnboardingActive]="isOnboardingActive"
+    [surveysLoaded]="surveysLoaded"
+    [isNotificationsOn]="isNotificationsOn"
+    [isNotificationLoading]="isNotificationLoading"
+    [activeSavedSearch]="activeSavedSearch"
+    (applyFilters)="applyFilters()"
+    (toggleMainFilters)="toggleMainFilters()"
+    (toggleSidebarFilters)="toggleSidebarFilters($event)"
+    (clearSavedFilter)="clearSavedFilter()"
+    (clearFilter)="clearFilter($event)"
+    (clearCollection)="clearCollection()"
+    (changeNorificationStatus)="changeNorificationStatus($event)"
+  ></app-survey-filters>
 </mat-toolbar>
 
 <app-mobile-menu></app-mobile-menu>
